--- src/pages/Chat/index.css
@@ -1,9 +1,0 @@
-.bg-blue {
-  background-color: #179f97 !important;
-  color: white !important;
-}
-
-.bg-user-color {
-  background-color: #cbcbcb !important;
-  text-align: right;
-}
